server/requirements.txt
flask
flask_restful
python-hcl2